FP SALES  ANALYSIS
This is a PowerBi project on camera sales for a period of time on a fictitious company known as FP. The aim of the project is to analyze the sales of the product over a given time an derive insights and solutions for the growth of the business.                                                                                                                                                                                     The data collected is in a c.s.v format which was imported into the PowerBi environment for further analysis. The images above shows the various processes the data went through from the cleaning to the new measures added,the modelling and the various relationships built to the final visualization of the data and insights gotten from the data. It equally shows the various steps applied and the various tables that is contained in the data given.